+++ docs/pages/web/callout.js
@@ -8,7 +8,20 @@ import Page from '../../docs-components/Page.js';
 import PageHeader from '../../docs-components/PageHeader.js';
 import QualityChecklist from '../../docs-components/QualityChecklist.js';
 import SandpackExample from '../../docs-components/SandpackExample.js';
+import accessibilityExample from '../../examples/callout/accessibilityExample.js';
+import actionsExample from '../../examples/callout/actionsExample.js';
+import dismissibleExample from '../../examples/callout/dismissibleExample.js';
+import dontStack from '../../examples/callout/dontStack.js';
+import dontUseForMarketing from '../../examples/callout/dontUseForMarketing.js';
+import localizationExample from '../../examples/callout/localizationExample.js';
 import main from '../../examples/callout/main.js';
+import placeAtTop from '../../examples/callout/placeAtTop.js';
+import productMessages from '../../examples/callout/productMessages.js';
+import variantError from '../../examples/callout/variantError.js';
+import variantInfo from '../../examples/callout/variantInfo.js';
+import variantRecommendation from '../../examples/callout/variantRecommendation.js';
+import variantSuccess from '../../examples/callout/variantSuccess.js';
+import variantWarning from '../../examples/callout/variantWarning.js';
 
 export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen |}): Node {
   return (
@@ -54,20 +67,13 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
             cardSize="lg"
             type="do"
             description="Use Callout for messages coming from the product or user interaction. Can be used in both Business and Pinner products."
-            defaultCode={`
-          <Callout
-            iconAccessibilityLabel="Error"
-            message="Your tag has errors, so information may be outdated. Fix your tag for the most accurate metrics."
-            primaryAction={{
-              accessibilityLabel: 'Fix Pinterest Tag',
-              href: "https://pinterest.com",
-              label: "Fix tag",
-              target: "blank",
-            }}
-            title="Pinterest tag needs attention"
-            type="error"
-          />
-        `}
+            sandpackExample={
+              <SandpackExample
+                name="Do - Use Callout for product messages"
+                code={productMessages}
+                hideEditor
+              />
+            }
           />
           <MainSection.Card
             cardSize="lg"
@@ -75,38 +81,13 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
             description={`
         Place Callout at the top of the page, under the primary navigation or page header when possible.
         `}
-            defaultCode={`
-          <Flex direction="column" gap={{ column: 4, row: 0 }}>
-            <Flex alignItems="center" justifyContent="start">
-              <Icon accessibilityLabel="" icon="pinterest" color="error" size={32}/>
-              <ButtonGroup>
-                <Button color="transparent" iconEnd="arrow-down" text="Business" />
-                <Button color="transparent" iconEnd="arrow-down" text="Create" />
-                <Button color="transparent" iconEnd="arrow-down" text="Analytics" />
-                <Button color="transparent" iconEnd="arrow-down" text="Ads" />
-              </ButtonGroup>
-            </Flex>
-
-            <Divider/>
-
-            <Box marginTop={4}>
-              <Callout
-                dismissButton={{
-                  accessibilityLabel: 'Dismiss this banner',
-                  onDismiss: () => {},
-                }}
-                iconAccessibilityLabel="Info"
-                message="It may take up to 10 minutes to automatically detect a newly installed tag. If you'd like to manually verify your tag, please click the Verify Tag button."
-                primaryAction={{
-                  accessibilityLabel: "Manually verify tag",
-                  label: "Verify Tag",
-                }}
-                title="We have not yet detected your tag"
-                type="info"
+            sandpackExample={
+              <SandpackExample
+                name="Do - Place Callout at the top of the page"
+                code={placeAtTop}
+                hideEditor
               />
-            </Box>
-          </Flex>
-        `}
+            }
           />
           <MainSection.Card
             cardSize="lg"
@@ -114,22 +95,13 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
             description={`
         Use Callouts for marketing new products or features. Use [Upsell](/web/upsell) instead.
         `}
-            defaultCode={`
-          <Callout
-            dismissButton={{
-              accessibilityLabel: 'Dismiss this banner',
-              onDismiss: () => {},
-            }}
-            iconAccessibilityLabel="Info"
-            message="Earn $60 of ads credit, and send $30 of ads credit to a friend"
-            primaryAction={{
-              accessibilityLabel: "Send ads invite",
-              label: "Send invite",
-            }}
-            title="Give $30, get $60 in ads credit"
-            type="info"
-          />
-        `}
+            sandpackExample={
+              <SandpackExample
+                name="Don't - Use Callouts for marketing new products or features"
+                code={dontUseForMarketing}
+                hideEditor
+              />
+            }
           />
           <MainSection.Card
             cardSize="lg"
@@ -137,51 +109,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
             description={`
         Stack Callouts. In the case that banners must be stacked, Callouts should come before Upsells.
         `}
-            defaultCode={`
-        <Flex direction="column" gap={{ column: 4, row: 0 }}>
-          <Flex alignItems="center" justifyContent="start">
-            <Icon accessibilityLabel="" icon="pinterest" color="error" size={32}/>
-            <ButtonGroup>
-              <Button color="transparent" iconEnd="arrow-down" text="Business" />
-              <Button color="transparent" iconEnd="arrow-down" text="Create" />
-              <Button color="transparent" iconEnd="arrow-down" text="Analytics" />
-              <Button color="transparent" iconEnd="arrow-down" text="Ads" />
-            </ButtonGroup>
-          </Flex>
-
-          <Divider/>
-
-          <Box marginTop={4}>
-            <Flex gap={{ column: 2, row: 0 }} direction="column">
-              <Upsell
-                imageData={{
-                  component: <Icon icon="send" accessibilityLabel="Send" color="default" size={32}/>
-                }}
-                message="Track ads conversion—sales, traffic and more—with the Pinterest tag"
-                primaryAction={{
-                  accessibilityLabel: "Claim ads credit now",
-                  label: "Claim now",
-                }}
-                title="So close! Finish installing your Pinterest tag, get $10 in ads credit"
-              />
-              <Callout
-                dismissButton={{
-                  accessibilityLabel: 'Dismiss this banner',
-                  onDismiss: () => {},
-                }}
-                iconAccessibilityLabel="Info"
-                message="It may take up to 10 minutes to automatically detect a newly installed tag. If you'd like to manually verify your tag, please click the Verify Tag button."
-                primaryAction={{
-                  accessibilityLabel: "Manually verify tag",
-                  label: "Verify Tag",
-                }}
-                title="We have not yet detected your tag"
-                type="info"
-              />
-            </Flex>
-          </Box>
-        </Flex>
-        `}
+            sandpackExample={
+              <SandpackExample name="Don't - Stack Callouts" code={dontStack} hideEditor />
+            }
           />
         </MainSection.Subsection>
       </MainSection>
@@ -199,30 +129,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  dismissButton={{
-    accessibilityLabel: 'Dismiss this banner',
-    onDismiss: () => {},
-  }}
-  iconAccessibilityLabel="Info"
-  message="Apply to the Verified Merchant Program"
-  primaryAction={{
-    accessibilityLabel: "Get started: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Get started",
-    target: "blank",
-  }}
-  secondaryAction={{
-    accessibilityLabel: "Learn more: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Learn more",
-    target: "blank",
-  }}
-  title="Your business account was created!"
-  type="info"
-/>
-        `}
+            sandpackExample={
+              <SandpackExample name="Callout labels" code={accessibilityExample} hideEditor />
+            }
           />
         </MainSection.Subsection>
       </AccessibilitySection>
@@ -234,30 +143,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         <MainSection.Subsection>
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  dismissButton={{
-    accessibilityLabel: 'Dismiss this banner',
-    onDismiss: () => {},
-  }}
-  iconAccessibilityLabel="Info"
-  message="Bewerben Sie sich beim Verified Merchant Program"
-  primaryAction={{
-    accessibilityLabel: "Loslegen: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Loslegen",
-    target: "blank",
-  }}
-  secondaryAction={{
-    accessibilityLabel: "Erfahren Sie mehr: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Erfahren Sie mehr",
-    target: "blank",
-  }}
-  title="Ihr Geschäftskonto wurde erstellt!"
-  type="info"
-/>
-        `}
+            sandpackExample={
+              <SandpackExample name="Callout localization" code={localizationExample} hideEditor />
+            }
           />
         </MainSection.Subsection>
       </MainSection>
@@ -269,30 +157,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  dismissButton={{
-    accessibilityLabel: 'Dismiss this banner',
-    onDismiss: () => {},
-  }}
-  iconAccessibilityLabel="Info"
-  message="Apply to the Verified Merchant Program"
-  primaryAction={{
-    accessibilityLabel: "Get started: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Get started",
-    target: "blank",
-  }}
-  secondaryAction={{
-    accessibilityLabel: "Learn more: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Learn more",
-    target: "blank",
-  }}
-  title="Your business account was created!"
-  type="info"
-/>
-`}
+            sandpackExample={
+              <SandpackExample name="Variants - Info" code={variantInfo} hideEditor />
+            }
           />
         </MainSection.Subsection>
 
@@ -302,24 +169,13 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  dismissButton={{
-    accessibilityLabel: 'Dismiss this banner',
-    onDismiss: () => {},
-  }}
-  iconAccessibilityLabel="Recommendation"
-  message="When you run ads on Pinterest, you'll find recommendations to improve them here."
-  primaryAction={{
-    accessibilityLabel: "Learn more: Ads on Pinterest",
-    href: "https://pinterest.com",
-    label: "Learn more",
-    target: "blank",
-  }}
-  title="Advertise with confidence!"
-  type="recommendation"
-/>
-`}
+            sandpackExample={
+              <SandpackExample
+                name="Variants - Recommendation"
+                code={variantRecommendation}
+                hideEditor
+              />
+            }
           />
         </MainSection.Subsection>
         <MainSection.Subsection
@@ -328,24 +184,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-  <Callout
-    dismissButton={{
-      accessibilityLabel: 'Dismiss this banner',
-      onDismiss: () => {},
-    }}
-    iconAccessibilityLabel="Success"
-    message="Keep it up by using recommendations to optimize your ad spend."
-    primaryAction={{
-      accessibilityLabel: "Get started: Ad recommendations",
-      href: "https://pinterest.com",
-      label: "Get started",
-      target: "blank",
-    }}
-    title="Your ads are doing great!"
-    type="success"
-  />
-`}
+            sandpackExample={
+              <SandpackExample name="Variants - Success" code={variantSuccess} hideEditor />
+            }
           />
         </MainSection.Subsection>
 
@@ -355,24 +196,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  dismissButton={{
-    accessibilityLabel: 'Dismiss warning',
-    onDismiss: () => {},
-  }}
-  iconAccessibilityLabel="Warning"
-  message="We have noticed that you have audiences in your advertiser account that have been used in an ad campaign. Pinterest will be deleting any unused audiences on May 30, 2020."
-  primaryAction={{
-    accessibilityLabel: "View unused audiences",
-    href: "https://pinterest.com",
-    label: "View audiences",
-    target: "blank",
-  }}
-  title="Unused audiences are going away"
-  type="warning"
-/>
-`}
+            sandpackExample={
+              <SandpackExample name="Variants - Warning" code={variantWarning} hideEditor />
+            }
           />
         </MainSection.Subsection>
 
@@ -382,20 +208,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  iconAccessibilityLabel="Error"
-  message="Your tag has errors, so information may be outdated. Fix your tag for the most accurate metrics."
-  primaryAction={{
-    accessibilityLabel: "Fix Pinterest tag",
-    href: "https://pinterest.com",
-    label: "Fix tag",
-    target: "blank",
-  }}
-  title="Pinterest tag needs attention"
-  type="error"
-/>
-`}
+            sandpackExample={
+              <SandpackExample name="Variants - Error" code={variantError} hideEditor />
+            }
           />
         </MainSection.Subsection>
 
@@ -414,98 +229,9 @@ export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-function Example(props) {
-  const [showModal, setShowModal] = React.useState(false);
-  const toggleModal = () => {
-    setShowModal((currState) => !currState);
-  };
-
-  return (
-    <Box marginStart={-1} marginEnd={-1}>
-      <Callout
-        dismissButton={{
-          accessibilityLabel: 'Dismiss banner',
-          onDismiss: () => {},
-        }}
-        iconAccessibilityLabel="Info"
-        message="Apply to the Verified Merchant Program"
-        primaryAction={{
-          accessibilityLabel: "Apply now: verified merchant program",
-          label: "Apply now",
-          onClick: toggleModal,
-        }}
-        secondaryAction={{
-          accessibilityLabel: "Learn more: Verified Merchant Program",
-          href: "https://help.pinterest.com/en/business/article/verified-merchant-program",
-          label: "Learn more",
-          target: "blank",
-        }}
-        title="Your business account was created!"
-        type="info"
-      />
-
-      {showModal && (
-        <Layer>
-          <Modal
-            accessibilityModalLabel="Apply for the Verified Merchant Program"
-            footer={
-              <Flex flex="grow" justifyContent="end">
-                  <ButtonGroup>
-                    <Button
-                      onClick={toggleModal}
-                      size="lg"
-                      text="Cancel"
-                    />
-                    <Button
-                      color="red"
-                      onClick={toggleModal}
-                      size="lg"
-                      text="Save"
-                    />
-                  </ButtonGroup>
-              </Flex>
+            sandpackExample={
+              <SandpackExample name="Callout actions" code={actionsExample} hideEditor />
             }
-            heading="Verified Merchant Program Application"
-            onDismiss={toggleModal}
-            size="md"
-          >
-            <Flex>
-              <Column span={12}>
-                <Box paddingY={2} paddingX={8} display="flex">
-                  <Column span={4}>
-                    <Label htmlFor="name">
-                      <Text align="start" weight="bold">
-                        Name
-                      </Text>
-                    </Label>
-                  </Column>
-                  <Column span={8}>
-                    <TextField id="name" onChange={() => {}} />
-                  </Column>
-                </Box>
-
-                <Box paddingY={2} paddingX={8} display="flex">
-                  <Column span={4}>
-                    <Label htmlFor="desc">
-                      <Text align="start" weight="bold">
-                        Business Description
-                      </Text>
-                    </Label>
-                  </Column>
-                  <Column span={8}>
-                    <TextArea id="desc" onChange={() => {}} />
-                  </Column>
-                </Box>
-              </Column>
-            </Flex>
-          </Modal>
-        </Layer>
-      )}
-    </Box>
-  );
-}
-        `}
           />
         </MainSection.Subsection>
 
@@ -521,30 +247,9 @@ function Example(props) {
         >
           <MainSection.Card
             cardSize="lg"
-            defaultCode={`
-<Callout
-  dismissButton={{
-    accessibilityLabel: 'Dismiss this banner',
-    onDismiss: () => {},
-  }}
-  iconAccessibilityLabel="Info"
-  message="Apply to the Verified Merchant Program"
-  primaryAction={{
-    accessibilityLabel: "Get started: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Get started",
-    target: "blank",
-  }}
-  secondaryAction={{
-    accessibilityLabel: "Learn more: Verified Merchant Program",
-    href: "https://pinterest.com",
-    label: "Learn more",
-    target: "blank",
-  }}
-  title="Your business account was created!"
-  type="info"
-/>
-      `}
+            sandpackExample={
+              <SandpackExample name="Dismissable Callout" code={dismissibleExample} hideEditor />
+            }
           />
         </MainSection.Subsection>
       </MainSection>
